+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - conftest: Default to using elf file in `bin/app.elf` if `--elfs_dir` is not specified
 - conftest: Targeted device must be specified, in doesn't default to all by default anymore
+- conftest: Use session scope for firmware, backend and navigator fixtures
 
 ### Fixed